Updated Binary Tree to minimize space waste
As mentioned in a previous commit, using the 2i+1 and 2i+2 formulas for
locating the child nodes doesn't really work for anything besides heaps.
Since the binary node was updated to have the needed pointers I modified
the binary tree to use that data, so it isn't space wasteful, its just
that the lookup isn't going to realistically be close to log(n).